# Source/BSN.IpTables.Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "IpTables API",
    "description": "\r\nIpTables.Api is the RESTful server which applies iptables rules to block or allow IP addresses on request from a remote client.\r\nIpTables.Api maintains the list of blocked IP addresses which may be updated on the fly from many sources.\r\nIpTables.Api also solves the problem of concurrent modifications to iptables since the requests are serialized.\r\n                            ",
    "termsOfService": "https://github.com/BSVN/IpTables.Api",
    "license": {
      "name": "MIT license",
      "url": "https://github.com/BSVN/IpTables.Api/blob/main/LICENSE"
    },
    "version": "1.0"
  },
  "paths": {
    "/api/v1/Home/List": {
      "get": {
        "tags": [
          "Home"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/IpTablesChainSetViewModelResponse"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/IpTablesChainSetViewModelResponse"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/IpTablesChainSetViewModelResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Home/Insert": {
      "put": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/RulesCommandServiceInsertRequest"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Home/Append": {
      "post": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "Chain",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.InterfaceName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.Protocol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.SourceIp",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.DestinationIp",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.SourcePort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.DestinationPort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.Jump",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Home/Delete": {
      "delete": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "Chain",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.InterfaceName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.Protocol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.SourceIp",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.DestinationIp",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.SourcePort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.DestinationPort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.Jump",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Home": {
      "delete": {
        "tags": [
          "Home"
        ],
        "parameters": [
          {
            "name": "Chain",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.InterfaceName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.Protocol",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.SourceIp",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.DestinationIp",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.SourcePort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.DestinationPort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Data.Jump",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "InvalidItem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IpTablesChainSetViewModel": {
        "type": "object",
        "properties": {
          "ipTablesChains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IpTablesChainViewModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IpTablesChainSetViewModelResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ResponseStatusCode"
              }
            ]
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IpTablesChainSetViewModel"
              }
            ],
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "invalidItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvalidItem"
            },
            "nullable": true
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "IpTablesChainViewModel": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "tableName": {
            "type": "string",
            "nullable": true
          },
          "ipVersion": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IpVersion"
              }
            ]
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IpTablesRuleViewModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IpTablesRuleViewModel": {
        "type": "object",
        "properties": {
          "interfaceName": {
            "type": "string",
            "nullable": true
          },
          "protocol": {
            "type": "string",
            "nullable": true
          },
          "sourceIp": {
            "type": "string",
            "nullable": true
          },
          "destinationIp": {
            "type": "string",
            "nullable": true
          },
          "sourcePort": {
            "type": "string",
            "nullable": true
          },
          "destinationPort": {
            "type": "string",
            "nullable": true
          },
          "target": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IpVersion": {
        "enum": [
          4,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "Response": {
        "type": "object",
        "properties": {
          "statusCode": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ResponseStatusCode"
              }
            ]
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "invalidItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvalidItem"
            },
            "nullable": true
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ResponseStatusCode": {
        "enum": [
          200,
          201,
          202,
          203,
          204,
          205,
          206,
          300,
          302,
          304,
          306,
          400,
          401,
          402,
          403,
          404,
          405,
          406,
          408,
          409,
          410,
          411,
          412,
          413,
          414,
          416,
          417,
          500,
          501,
          502,
          503,
          504
        ],
        "type": "integer",
        "format": "int32"
      },
      "RulesCommandServiceInsertRequest": {
        "type": "object",
        "additionalProperties": false
      }
    }
  }
}